+++ src/components/ConfirmDeleteQuestionModal.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 import ConfirmModal from './ConfirmModal'
 
 const ConfirmDeleteQuestionModal = props => (
@@ -11,6 +12,10 @@ const ConfirmDeleteQuestionModal = props => (
   />
 )
 
-ConfirmDeleteQuestionModal.propTypes = ConfirmModal.propTypes
+ConfirmDeleteQuestionModal.propTypes = {
+  isOpen: PropTypes.bool.isRequired,
+  toggle: PropTypes.func.isRequired,
+  confirm: PropTypes.func.isRequired,
+}
 
 export default ConfirmDeleteQuestionModal