+++ src/components/ConfirmDeleteQueueModal.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 import ConfirmModal from './ConfirmModal'
 
 const ConfirmDeleteQueueModal = props => (
@@ -11,6 +12,10 @@ const ConfirmDeleteQueueModal = props => (
   />
 )
 
-ConfirmDeleteQueueModal.propTypes = ConfirmModal.propTypes
+ConfirmDeleteQueueModal.propTypes = {
+  isOpen: PropTypes.bool.isRequired,
+  toggle: PropTypes.func.isRequired,
+  confirm: PropTypes.func.isRequired,
+}
 
 export default ConfirmDeleteQueueModal